# swagger.yaml
openapi: 3.0.0
info:
  title: Node MySQL API
  version: 1.0.0
  description: Node.js + MySQL - API with email sign-up, verification, authentication and forgot password
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
  /accounts/authenticate:
    post:
      tags: [default]
      summary: Authenticate account credentials and return a JWT token and a cookie with a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK

  /accounts/refresh-token:
    post:
      tags: [default]
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      responses:
        '200':
          description: OK

  /accounts/revoke-token:
    post:
      tags: [default]
      summary: Revoke a refresh token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
      responses:
        '200':
          description: OK

  /accounts/register:
    post:
      tags: [default]
      summary: Register a new user account and send a verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string }
                password: { type: string }
                confirmPassword: { type: string }
                acceptTerms: { type: boolean }
      responses:
        '200':
          description: OK

  /accounts/verify-email:
    post:
      tags: [default]
      summary: Verify a new account with a verification token received by email after registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
      responses:
        '200':
          description: OK

  /accounts/forgot-password:
    post:
      tags: [default]
      summary: Submit email address to reset the password on an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200':
          description: OK

  /accounts/reset-password:
    post:
      tags: [default]
      summary: Reset the password for an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                password: { type: string }
                confirmPassword: { type: string }
      responses:
        '200':
          description: OK

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []